[#13][#56] split useInput and useCheckbox
