+++ src/lib/Input/useInput.ts
@@ -1,20 +1,14 @@
 import React, { useState, ChangeEvent } from 'react'
 
-interface IInputOptions {
-	isCheckbox: boolean;
-}
-
-type S = (string | number | boolean)
-type E = ChangeEvent<any>
+type S = (string | number)
+type E = ChangeEvent<HTMLInputElement | HTMLSelectElement>
 type F = (e: E) => void;
 
-const useInput = (props?: IInputOptions): [S, F] => {
+const useInput = (): [S, F] => {
 	const [value, setValue] = useState(null)
-	const { isCheckbox } = props || {}
 
 	const handleValue = (e: E): void => {
-		const value = isCheckbox === undefined || !isCheckbox ? e.currentTarget.value : e.currentTarget.checked
-		setValue(value)
+		setValue(e.currentTarget.value)
 	}
 
 	return [value, handleValue]
